requirements.txt
google-generativeai
bs4
python-dotenv